refactor(core,gc): split bloated modules into impl + tests files
This branch grew two files past the cognitive-load threshold:
* `src/core/virtualenv.rs`: 431 → 990 lines (+559) for the
  last_used / atomic-write / touch surface and its regression
  suite.
* `src/cli/commands/gc.rs`: 803 → 1376 lines (+573) for the
  Stale reason, scan/recheck stale flow, path-traversal guard
  tests, and JSON-shape pins.

Convert both single-file modules to directory modules with the
impl in `mod.rs` and the test suite extracted to a sibling
`tests.rs`. Pure refactor — line count balances exactly (1633
insertions vs 1633 deletions), no functional change, all 776 unit +
45 integration + 23 doctest still pass; clippy clean.

After:
* `src/core/virtualenv/mod.rs`: 397 lines (impl + `mod tests;`)
* `src/core/virtualenv/tests.rs`: 593 lines (tests, de-indented)
* `src/cli/commands/gc/mod.rs`: 735 lines
* `src/cli/commands/gc/tests.rs`: 641 lines

Why test extraction rather than a deeper logical split (types/scan/
remove for gc, etc.): same-module test visibility stays intact via
`use super::*;` with no `pub(super)` surgery on private types, so the
diff stays a mechanical move that any future reviewer can verify by
checking line counts and running the test suite. A deeper architectural
split inside `gc/` (types.rs / scan.rs / remove.rs / mod.rs) is a
reasonable next step but is left as a follow-up because it would
require touching every test that constructs `OrphanEnv` / `EnvRecord`
to import re-exported types, and would mix structural refactor with
test reorganization in the same commit.
