--- src/core/virtualenv.rs
@@ -1,990 +1,0 @@
-//! Virtual environment service
-
-use std::fs;
-use std::io::Write;
-use std::path::{Path, PathBuf};
-
-use chrono::{DateTime, Utc};
-use tracing::warn;
-
-use crate::core::Metadata;
-use crate::error::{Result, ScoopError};
-use crate::paths;
-use crate::uv::UvClient;
-use crate::validate;
-
-/// Information about a virtual environment.
-///
-/// `#[non_exhaustive]` so future field additions (the v2 plan calls
-/// for at least one more metadata-derived field down the line) aren't
-/// a breaking change for external `scoop-uv` library consumers. All
-/// in-tree construction sites use struct-literal syntax and live in
-/// `cfg(test)` or this crate's command handlers, so the attribute
-/// only affects downstream consumers — which currently must build
-/// `VirtualenvInfo` via field-by-field literal anyway.
-#[derive(Debug, Clone)]
-#[non_exhaustive]
-pub struct VirtualenvInfo {
-    /// Name of the environment
-    pub name: String,
-    /// Path to the environment
-    pub path: PathBuf,
-    /// Python version (if metadata exists)
-    pub python_version: Option<String>,
-    /// Creation timestamp from metadata, if present. Populated by
-    /// [`VirtualenvService::list`] so callers (like `list --sort=created`)
-    /// don't need a second `read_metadata` round trip per env.
-    pub created_at: Option<DateTime<Utc>>,
-    /// Last-used timestamp from metadata, if present. Same rationale:
-    /// list-time sort can use it without re-reading the JSON file.
-    pub last_used: Option<DateTime<Utc>>,
-}
-
-/// Service for managing virtual environments
-pub struct VirtualenvService {
-    uv: UvClient,
-}
-
-impl VirtualenvService {
-    /// Create a new service with the given uv client
-    pub fn new(uv: UvClient) -> Self {
-        Self { uv }
-    }
-
-    /// Create a new service, finding uv automatically
-    pub fn auto() -> Result<Self> {
-        Ok(Self::new(UvClient::new()?))
-    }
-
-    /// List all virtual environments
-    pub fn list(&self) -> Result<Vec<VirtualenvInfo>> {
-        let venvs_dir = paths::virtualenvs_dir()?;
-
-        if !venvs_dir.exists() {
-            return Ok(Vec::new());
-        }
-
-        let mut envs = Vec::new();
-
-        for entry in fs::read_dir(&venvs_dir)? {
-            // Per-entry tolerance — transient IO errors on a single entry
-            // shouldn't hide the rest of the directory from callers.
-            let entry = match entry {
-                Ok(e) => e,
-                Err(_) => continue,
-            };
-            // Reject symlinks via file_type() (no traversal) instead of
-            // path.is_dir() (which follows symlinks). A symlink under
-            // virtualenvs/ would otherwise be enumerated as a normal env,
-            // and downstream commands like `scoop verify` would exec the
-            // target's bin/python — arbitrary execution under the user's
-            // UID. This is the same hardening gc::scan_orphan_envs does;
-            // doing it here makes every caller of list() consistent.
-            let ft = match entry.file_type() {
-                Ok(t) => t,
-                Err(_) => continue,
-            };
-            if !ft.is_dir() || ft.is_symlink() {
-                continue;
-            }
-            let path = entry.path();
-            if let Some(name) = path.file_name().and_then(|n| n.to_str()) {
-                // Single metadata read per entry: we extract every field
-                // the new VirtualenvInfo wants (python_version + the two
-                // timestamps) in one shot, so callers that sort by
-                // created_at / last_used don't drive a second pass over
-                // every metadata file. Cheap when corrupt — `None` is a
-                // valid bucket-end value for sort, and the legacy passive
-                // contract (display "-"/"never") is preserved.
-                let metadata = self.read_metadata(&path);
-                let (python_version, created_at, last_used) = match metadata {
-                    Some(m) => (Some(m.python_version), Some(m.created_at), m.last_used),
-                    None => (None, None, None),
-                };
-                envs.push(VirtualenvInfo {
-                    name: name.to_string(),
-                    path: path.clone(),
-                    python_version,
-                    created_at,
-                    last_used,
-                });
-            }
-        }
-
-        envs.sort_by(|a, b| a.name.cmp(&b.name));
-        Ok(envs)
-    }
-
-    /// Create a new virtual environment
-    pub fn create(&self, name: &str, python_version: &str) -> Result<PathBuf> {
-        self.create_inner(name, python_version, None)
-    }
-
-    /// Create a new virtual environment using a specific Python executable path.
-    ///
-    /// The `python_path` is passed directly to uv's `--python` flag, which
-    /// accepts both version strings and paths. The `python_version` should be
-    /// the detected version string from the binary. The canonical path is stored
-    /// in metadata.
-    pub fn create_with_python_path(
-        &self,
-        name: &str,
-        python_version: &str,
-        python_path: &Path,
-    ) -> Result<PathBuf> {
-        self.create_inner(
-            name,
-            &python_path.display().to_string(),
-            Some((python_version, python_path)),
-        )
-    }
-
-    /// Internal create implementation shared by both create methods.
-    fn create_inner(
-        &self,
-        name: &str,
-        uv_python_arg: &str,
-        python_path_info: Option<(&str, &Path)>,
-    ) -> Result<PathBuf> {
-        validate::validate_env_name(name)?;
-
-        let path = paths::virtualenv_path(name)?;
-
-        if path.exists() {
-            return Err(ScoopError::VirtualenvExists {
-                name: name.to_string(),
-            });
-        }
-
-        // Ensure parent directory exists
-        if let Some(parent) = path.parent() {
-            fs::create_dir_all(parent)?;
-        }
-
-        // Create the virtual environment
-        self.uv.create_venv(&path, uv_python_arg)?;
-
-        // Write metadata
-        let uv_version = self.uv.version().ok();
-        // Resolve actual version: prefer pyvenv.cfg (handles specifiers like cpython@3.12),
-        // then explicit python_path version, then fall back to the raw uv arg.
-        let actual_version = super::parse_pyvenv_version(&path)
-            .or_else(|| python_path_info.map(|(ver, _)| ver.to_string()))
-            .unwrap_or_else(|| uv_python_arg.to_string());
-        let mut metadata = Metadata::new(name.to_string(), actual_version, uv_version);
-
-        if let Some((_, pp)) = python_path_info {
-            metadata = metadata.with_python_path(pp.display().to_string());
-        }
-
-        self.write_metadata_atomic(&path, &metadata)?;
-
-        Ok(path)
-    }
-
-    /// Delete a virtual environment.
-    ///
-    /// Validates `name` internally before touching the filesystem.
-    /// `PathBuf::join` does not block `..` and silently replaces the
-    /// base when the right side is absolute, so `delete("/tmp/x")`
-    /// against an unvalidated name would happily walk into and remove
-    /// the target directory under the user's UID. The validation guard
-    /// here is the trust boundary — CLI handlers no longer have to
-    /// remember to call `validate_env_name` themselves for safety.
-    pub fn delete(&self, name: &str) -> Result<()> {
-        validate::validate_env_name(name)?;
-        let path = paths::virtualenv_path(name)?;
-
-        if !path.exists() {
-            return Err(ScoopError::VirtualenvNotFound {
-                name: name.to_string(),
-            });
-        }
-
-        fs::remove_dir_all(&path)?;
-        Ok(())
-    }
-
-    /// Check whether a Python version matching `version` is already installed
-    /// via uv. Thin pass-through to [`UvClient::find_python`] so command
-    /// handlers don't need direct access to the private `uv` field.
-    pub fn is_python_installed(&self, version: &str) -> Result<bool> {
-        Ok(self.uv.find_python(version)?.is_some())
-    }
-
-    /// Install a Python version through uv. Thin pass-through that lets command
-    /// handlers stay decoupled from the private `uv` field.
-    pub fn install_python(&self, version: &str) -> Result<()> {
-        self.uv.install_python(version)
-    }
-
-    /// Install Python packages into the env via uv. Thin pass-through so the
-    /// sync handler doesn't need direct access to the private `uv` field.
-    pub fn pip_install(&self, venv_path: &Path, packages: &[String]) -> Result<()> {
-        self.uv.pip_install(venv_path, packages)
-    }
-
-    /// Check if a virtual environment exists.
-    ///
-    /// Validates `name` internally — see [`Self::delete`] for the path
-    /// traversal rationale. A `false` return for an invalid name would
-    /// hide the bug instead of surfacing it, so we error out.
-    pub fn exists(&self, name: &str) -> Result<bool> {
-        validate::validate_env_name(name)?;
-        let path = paths::virtualenv_path(name)?;
-        Ok(path.exists())
-    }
-
-    /// Get the path to a virtual environment.
-    ///
-    /// Validates `name` internally — see [`Self::delete`] for the path
-    /// traversal rationale.
-    pub fn get_path(&self, name: &str) -> Result<PathBuf> {
-        validate::validate_env_name(name)?;
-        let path = paths::virtualenv_path(name)?;
-        if !path.exists() {
-            return Err(ScoopError::VirtualenvNotFound {
-                name: name.to_string(),
-            });
-        }
-        Ok(path)
-    }
-
-    /// Read metadata from a virtual environment.
-    ///
-    /// Returns `None` for both "file missing" and "file corrupt" — callers
-    /// that only need a best-effort view (e.g. `list`, `info`, `status`)
-    /// can keep using this. Anything that needs to *act* on the distinction
-    /// (gc classification, touch on activation) must use
-    /// [`Self::read_metadata_result`] instead.
-    pub fn read_metadata(&self, path: &Path) -> Option<Metadata> {
-        self.read_metadata_result(path).ok().flatten()
-    }
-
-    /// Read metadata distinguishing missing from corrupt.
-    ///
-    /// - `Ok(Some(m))` — file present, parsed cleanly
-    /// - `Ok(None)`    — file does not exist (legitimate "no metadata")
-    /// - `Err(e)`      — file present but unreadable / unparseable
-    ///
-    /// This split exists so [`Self::touch_metadata_best_effort`] can refuse
-    /// to overwrite a corrupt file (which would silently destroy the user's
-    /// only forensic trace of the corruption), while still updating files
-    /// that are merely absent.
-    pub fn read_metadata_result(&self, path: &Path) -> Result<Option<Metadata>> {
-        let metadata_path = path.join(Metadata::FILE_NAME);
-        let content = match fs::read_to_string(&metadata_path) {
-            Ok(c) => c,
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => return Ok(None),
-            Err(e) => return Err(ScoopError::Io(e)),
-        };
-        let parsed: Metadata = serde_json::from_str(&content)?;
-        Ok(Some(parsed))
-    }
-
-    /// Write metadata using an atomic replace: write to a sibling tempfile
-    /// in the same directory, then rename over the target.
-    ///
-    /// Same-directory tempfile is required so the rename stays on one
-    /// filesystem — cross-device rename via `fs::rename` / tempfile fails
-    /// with `EXDEV` rather than degrading to copy+delete, so a sibling
-    /// tempfile is what makes the rename viable at all.
-    ///
-    /// The rename itself is atomic in the visible-state sense: on Unix
-    /// it's `rename(2)`; on Windows tempfile uses `MoveFileExW` with
-    /// `MOVEFILE_REPLACE_EXISTING`. A *normal process crash* mid-write
-    /// therefore leaves either the old file intact or the new file
-    /// in place — readers never observe a half-written file.
-    ///
-    /// This is NOT a full power-loss durability promise. We don't `fsync`
-    /// the file or the parent directory: this is best-effort metadata
-    /// (timestamps for display + gc heuristics), and `sync_all` on every
-    /// auto-activation would put a disk flush on the `cd` hot path. If a
-    /// power loss hits between the rename and the cache flush, the
-    /// metadata may roll back to its previous state. We accept that.
-    pub fn write_metadata_atomic(&self, path: &Path, metadata: &Metadata) -> Result<()> {
-        let metadata_path = path.join(Metadata::FILE_NAME);
-        let dir = metadata_path.parent().ok_or_else(|| {
-            ScoopError::Io(std::io::Error::new(
-                std::io::ErrorKind::InvalidInput,
-                "metadata path has no parent",
-            ))
-        })?;
-
-        let content = serde_json::to_string_pretty(metadata)?;
-        let mut tmp = tempfile::NamedTempFile::new_in(dir)?;
-        tmp.write_all(content.as_bytes())?;
-        tmp.persist(&metadata_path)
-            .map_err(|e| ScoopError::Io(e.error))?;
-        Ok(())
-    }
-
-    /// Touch an env's `last_used` to *now* (wall clock at write time),
-    /// best-effort.
-    ///
-    /// Production entry point. The timestamp is sampled inside this
-    /// function so two racing callers (e.g. `scoop activate myenv` from
-    /// two shells) write *current* values instead of stale "now at
-    /// caller-time" values. This narrows — but does not eliminate — the
-    /// regression window between racing touches.
-    ///
-    /// Concurrency contract: **last-writer-wins**. The read→mutate→write
-    /// sequence is not locked, so under contention the file's final
-    /// timestamp reflects whichever process committed the rename last,
-    /// not necessarily the most recent wall-clock instant. Acceptable for
-    /// display ("Last used: 2 hours ago") and gc heuristics at the
-    /// day/week granularity Step 5 will offer; not acceptable for
-    /// anything that needs strict ordering.
-    ///
-    /// Never returns an error: activation must not be blocked by metadata
-    /// I/O failure. Three documented behaviors:
-    ///
-    /// 1. **Missing metadata** — skipped silently (legacy env that was
-    ///    never `scoop create`d via this binary). No file is created.
-    /// 2. **Corrupt metadata** — `warn!` logged and the file left
-    ///    untouched. Overwriting would destroy the user's only on-disk
-    ///    evidence of the corruption. (Warning is observability sugar,
-    ///    not a tested contract.)
-    /// 3. **Healthy metadata** — `last_used` updated via atomic replace.
-    pub fn touch_metadata_best_effort(&self, env_name: &str) {
-        self.touch_metadata_at(env_name, Utc::now());
-    }
-
-    /// Test seam: same behavior as [`Self::touch_metadata_best_effort`]
-    /// but with an explicit timestamp so tests can pin a deterministic
-    /// `last_used` value. Not for production callers — using it from
-    /// activate/run/shell would re-introduce the caller-stale timestamp
-    /// race that the public entry point exists to narrow.
-    pub(crate) fn touch_metadata_at(&self, env_name: &str, now: DateTime<Utc>) {
-        // Validation guard — see Self::delete for the path traversal
-        // rationale. Touch is best-effort, so we warn instead of
-        // returning; the public callers (activate/run/shell) already
-        // validate first so this is purely defense-in-depth against
-        // future internal callers.
-        if let Err(e) = validate::validate_env_name(env_name) {
-            warn!("touch_metadata: rejecting invalid env name {env_name:?}: {e}");
-            return;
-        }
-        let path = match paths::virtualenv_path(env_name) {
-            Ok(p) => p,
-            Err(e) => {
-                warn!("touch_metadata: cannot resolve path for {env_name}: {e}");
-                return;
-            }
-        };
-
-        match self.read_metadata_result(&path) {
-            Ok(Some(mut meta)) => {
-                meta.touch(now);
-                if let Err(e) = self.write_metadata_atomic(&path, &meta) {
-                    warn!("touch_metadata: atomic write failed for {env_name}: {e}");
-                }
-            }
-            Ok(None) => {
-                // Legacy env with no metadata file. Nothing to touch and
-                // we deliberately do NOT synthesize one — that would lie
-                // about created_at/created_by.
-            }
-            Err(e) => {
-                warn!("touch_metadata: refusing to overwrite corrupt metadata for {env_name}: {e}");
-            }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::test_utils::{create_mock_venv, with_temp_scoop_home};
-    use serial_test::serial;
-
-    /// Helper to get VirtualenvService, skipping test if uv not available.
-    /// Returns None if uv is not installed, allowing graceful test skip.
-    fn get_service() -> Option<VirtualenvService> {
-        crate::uv::UvClient::new().ok().map(VirtualenvService::new)
-    }
-
-    /// Macro to skip test if uv is not available.
-    /// This makes the skip explicit in test output.
-    macro_rules! require_uv {
-        () => {
-            match get_service() {
-                Some(service) => service,
-                None => {
-                    eprintln!("SKIPPED: uv not installed");
-                    return;
-                }
-            }
-        };
-    }
-
-    #[test]
-    fn test_virtualenv_info_struct() {
-        let info = VirtualenvInfo {
-            name: "testenv".to_string(),
-            path: PathBuf::from("/path/to/env"),
-            python_version: Some("3.12".to_string()),
-            created_at: None,
-            last_used: None,
-        };
-
-        assert_eq!(info.name, "testenv");
-        assert_eq!(info.path, PathBuf::from("/path/to/env"));
-        assert_eq!(info.python_version, Some("3.12".to_string()));
-    }
-
-    #[test]
-    #[serial]
-    fn test_list_empty_when_no_venvs_dir() {
-        with_temp_scoop_home(|_temp_dir| {
-            let service = require_uv!();
-            let result = service.list().unwrap();
-            assert!(result.is_empty());
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_list_returns_envs_sorted() {
-        with_temp_scoop_home(|temp_dir| {
-            // Arrange: Create mock venvs in reverse alphabetical order
-            create_mock_venv(temp_dir, "zeta", Some("3.11"));
-            create_mock_venv(temp_dir, "alpha", Some("3.12"));
-            create_mock_venv(temp_dir, "beta", None);
-
-            // Act
-            let service = require_uv!();
-            let envs = service.list().unwrap();
-
-            // Assert
-            assert_eq!(envs.len(), 3);
-            assert_eq!(envs[0].name, "alpha");
-            assert_eq!(envs[1].name, "beta");
-            assert_eq!(envs[2].name, "zeta");
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_list_reads_python_version_from_metadata() {
-        with_temp_scoop_home(|temp_dir| {
-            create_mock_venv(temp_dir, "withversion", Some("3.12.1"));
-            create_mock_venv(temp_dir, "noversion", None);
-
-            let service = require_uv!();
-            let envs = service.list().unwrap();
-
-            let with_ver = envs.iter().find(|e| e.name == "withversion").unwrap();
-            let no_ver = envs.iter().find(|e| e.name == "noversion").unwrap();
-
-            assert_eq!(with_ver.python_version, Some("3.12.1".to_string()));
-            assert_eq!(no_ver.python_version, None);
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_exists_returns_false_for_nonexistent() {
-        with_temp_scoop_home(|_temp_dir| {
-            let service = require_uv!();
-            assert!(!service.exists("nonexistent").unwrap());
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_exists_returns_true_for_existing() {
-        with_temp_scoop_home(|temp_dir| {
-            create_mock_venv(temp_dir, "exists", None);
-
-            let service = require_uv!();
-            assert!(service.exists("exists").unwrap());
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_get_path_returns_error_for_nonexistent() {
-        with_temp_scoop_home(|_temp_dir| {
-            let service = require_uv!();
-            let result = service.get_path("nonexistent");
-
-            assert!(result.is_err());
-            let err = result.unwrap_err();
-            assert!(matches!(err, ScoopError::VirtualenvNotFound { .. }));
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_get_path_returns_path_for_existing() {
-        with_temp_scoop_home(|temp_dir| {
-            create_mock_venv(temp_dir, "myenv", None);
-
-            let service = require_uv!();
-            let path = service.get_path("myenv").unwrap();
-            assert!(path.ends_with("myenv"));
-            assert!(path.exists());
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_delete_removes_directory() {
-        with_temp_scoop_home(|temp_dir| {
-            create_mock_venv(temp_dir, "todelete", Some("3.12"));
-            let venv_path = temp_dir.path().join("virtualenvs").join("todelete");
-            assert!(venv_path.exists());
-
-            let service = require_uv!();
-            service.delete("todelete").unwrap();
-            assert!(!venv_path.exists());
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_delete_returns_error_for_nonexistent() {
-        with_temp_scoop_home(|temp_dir| {
-            // Arrange: Create virtualenvs dir but not the specific venv
-            fs::create_dir_all(temp_dir.path().join("virtualenvs")).unwrap();
-
-            // Act
-            let service = require_uv!();
-            let result = service.delete("nonexistent");
-
-            // Assert
-            assert!(result.is_err());
-            let err = result.unwrap_err();
-            assert!(matches!(err, ScoopError::VirtualenvNotFound { .. }));
-        });
-    }
-
-    // ==========================================================================
-    // Path-traversal regression suite — every public name-taking method
-    // on VirtualenvService must reject hostile names BEFORE the path is
-    // joined. `PathBuf::join("/tmp/x")` returns `/tmp/x` (base discarded
-    // when right side is absolute), and `join("../foo")` doesn't strip
-    // the `..`; without the internal validate guard, a CLI handler that
-    // forgot to validate could hand attacker input straight to
-    // `remove_dir_all`. The guard is the trust boundary.
-    // ==========================================================================
-
-    fn bad_names() -> &'static [&'static str] {
-        &["/tmp/evil", "../escape", "../../tmp/evil", "/", "."]
-    }
-
-    #[test]
-    #[serial]
-    fn test_delete_rejects_path_traversal() {
-        with_temp_scoop_home(|temp_dir| {
-            fs::create_dir_all(temp_dir.path().join("virtualenvs")).unwrap();
-            let service = require_uv!();
-            for bad in bad_names() {
-                let err = service.delete(bad).unwrap_err();
-                assert!(
-                    matches!(err, ScoopError::InvalidEnvName { .. }),
-                    "delete({bad:?}) should reject as InvalidEnvName, got: {err:?}"
-                );
-            }
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_exists_rejects_path_traversal() {
-        with_temp_scoop_home(|temp_dir| {
-            fs::create_dir_all(temp_dir.path().join("virtualenvs")).unwrap();
-            let service = require_uv!();
-            for bad in bad_names() {
-                let err = service.exists(bad).unwrap_err();
-                assert!(
-                    matches!(err, ScoopError::InvalidEnvName { .. }),
-                    "exists({bad:?}): {err:?}"
-                );
-            }
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_get_path_rejects_path_traversal() {
-        with_temp_scoop_home(|temp_dir| {
-            fs::create_dir_all(temp_dir.path().join("virtualenvs")).unwrap();
-            let service = require_uv!();
-            for bad in bad_names() {
-                let err = service.get_path(bad).unwrap_err();
-                assert!(
-                    matches!(err, ScoopError::InvalidEnvName { .. }),
-                    "get_path({bad:?}): {err:?}"
-                );
-            }
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_touch_metadata_at_blocks_relative_path_escape() {
-        // Real teeth on the validation guard: plant a metadata file at
-        // the *exact* location `"../escape"` would resolve to without
-        // the guard (`<SCOOP_HOME>/virtualenvs/../escape` =
-        // `<SCOOP_HOME>/escape`). Without the guard, touch would read
-        // this file, mutate last_used, and persist it back — so the
-        // captured serialized bytes would change. With the guard, the
-        // file is never opened and the bytes stay identical.
-        //
-        // We use raw bytes (not the live Metadata struct) for the
-        // canary so the assertion is byte-for-byte; a fresh re-serialize
-        // would re-order fields or rewrite whitespace identically and
-        // mask the read-modify-write that this test exists to catch.
-        with_temp_scoop_home(|temp_dir| {
-            fs::create_dir_all(temp_dir.path().join("virtualenvs")).unwrap();
-            let escape_dir = temp_dir.path().join("escape");
-            fs::create_dir_all(&escape_dir).unwrap();
-            // Hand-rolled JSON that's parseable as Metadata (so the
-            // hypothetical unguarded read would succeed) but written
-            // with a recognizable last_used the test can distinguish
-            // from the new write.
-            let canary_path = escape_dir.join(".scoop-metadata.json");
-            let canary_bytes = b"{\
-                \"name\":\"escape\",\
-                \"python_version\":\"3.12\",\
-                \"created_at\":\"2024-01-01T00:00:00Z\",\
-                \"created_by\":\"scoop canary\",\
-                \"uv_version\":null,\
-                \"last_used\":\"1999-12-31T23:59:59Z\"\
-            }";
-            fs::write(&canary_path, canary_bytes).unwrap();
-
-            let service = require_uv!();
-            let new_now = "2026-06-02T12:00:00Z".parse().unwrap();
-            service.touch_metadata_at("../escape", new_now);
-
-            // Byte-for-byte unchanged ⇒ the validation guard short-
-            // circuited before read_metadata_result was called.
-            // Without the guard, last_used would have flipped from
-            // 1999-... to 2026-06-02 (and serde_json::to_string_pretty
-            // would have reformatted whitespace too).
-            let actual = fs::read(&canary_path).unwrap();
-            assert_eq!(
-                actual, canary_bytes,
-                "validation guard must short-circuit before the escape \
-                 location's metadata is read or overwritten",
-            );
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_touch_metadata_at_blocks_absolute_path_escape() {
-        // Companion to the relative-escape test: absolute paths bypass
-        // the base entirely via `PathBuf::join`. Plant a canary in a
-        // separate tempdir, hand `touch_metadata_at` that absolute
-        // path, and confirm the canary's bytes are unchanged.
-        with_temp_scoop_home(|temp_dir| {
-            fs::create_dir_all(temp_dir.path().join("virtualenvs")).unwrap();
-            let outside = tempfile::TempDir::new().unwrap();
-            let canary_path = outside.path().join(".scoop-metadata.json");
-            let canary_bytes = b"{\
-                \"name\":\"outside\",\
-                \"python_version\":\"3.12\",\
-                \"created_at\":\"2024-01-01T00:00:00Z\",\
-                \"created_by\":\"scoop canary\",\
-                \"uv_version\":null,\
-                \"last_used\":\"1999-12-31T23:59:59Z\"\
-            }";
-            fs::write(&canary_path, canary_bytes).unwrap();
-
-            let service = require_uv!();
-            let new_now = "2026-06-02T12:00:00Z".parse().unwrap();
-            // Pass the absolute path to outside/ as the "env name".
-            // Without the validation guard, paths::virtualenv_path
-            // would join this onto virtualenvs_dir and discard the
-            // base (Rust's PathBuf::join semantics), landing exactly
-            // on the canary.
-            let bad_name = outside.path().to_str().unwrap();
-            service.touch_metadata_at(bad_name, new_now);
-
-            let actual = fs::read(&canary_path).unwrap();
-            assert_eq!(
-                actual, canary_bytes,
-                "validation guard must reject absolute path before the \
-                 canary is read or overwritten",
-            );
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_list_ignores_files() {
-        with_temp_scoop_home(|temp_dir| {
-            let venvs_dir = temp_dir.path().join("virtualenvs");
-            fs::create_dir_all(&venvs_dir).unwrap();
-
-            // Create a file (not directory) - should be ignored
-            fs::write(venvs_dir.join("notadir"), "test").unwrap();
-            // Create a real venv directory
-            create_mock_venv(temp_dir, "realenv", None);
-
-            let service = require_uv!();
-            let envs = service.list().unwrap();
-
-            assert_eq!(envs.len(), 1);
-            assert_eq!(envs[0].name, "realenv");
-        });
-    }
-
-    // ==========================================================================
-    // last_used / atomic write / touch_metadata_best_effort
-    // ==========================================================================
-
-    fn seed_metadata_file(env_path: &Path, json: &str) {
-        fs::create_dir_all(env_path).unwrap();
-        fs::write(env_path.join(Metadata::FILE_NAME), json).unwrap();
-    }
-
-    #[test]
-    #[serial]
-    fn test_read_metadata_result_distinguishes_missing_from_corrupt() {
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_dir = temp_dir.path().join("virtualenvs").join("subject");
-            fs::create_dir_all(&env_dir).unwrap();
-
-            // No file → Ok(None). Distinct from "corrupt" because we
-            // shouldn't warn or refuse anything for a missing file.
-            assert!(service.read_metadata_result(&env_dir).unwrap().is_none());
-
-            // Corrupt JSON → Err(Json(..)). Caller needs this signal to
-            // decide whether to overwrite (no) or warn (yes).
-            fs::write(env_dir.join(Metadata::FILE_NAME), "{ not json").unwrap();
-            let err = service.read_metadata_result(&env_dir).unwrap_err();
-            assert!(
-                matches!(err, ScoopError::Json(_)),
-                "corrupt JSON must yield ScoopError::Json, got: {err:?}"
-            );
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_touch_metadata_best_effort_updates_only_last_used() {
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_path = temp_dir.path().join("virtualenvs").join("touched");
-            let seed = r#"{
-                "name": "touched",
-                "python_version": "3.12.1",
-                "created_at": "2024-01-15T10:30:00Z",
-                "created_by": "scoop 0.5.0",
-                "uv_version": "0.4.0"
-            }"#;
-            seed_metadata_file(&env_path, seed);
-
-            let now = "2026-06-02T12:00:00Z".parse::<DateTime<Utc>>().unwrap();
-            service.touch_metadata_at("touched", now);
-
-            let after = service
-                .read_metadata_result(&env_path)
-                .unwrap()
-                .expect("file present after touch");
-            assert_eq!(after.last_used, Some(now));
-            // Provenance fields must survive a touch — they describe how
-            // the env was created and a touch is not a re-creation.
-            assert_eq!(after.name, "touched");
-            assert_eq!(after.python_version, "3.12.1");
-            assert_eq!(after.created_by, "scoop 0.5.0");
-            assert_eq!(after.uv_version, Some("0.4.0".to_string()));
-            let expected_created_at = "2024-01-15T10:30:00Z".parse::<DateTime<Utc>>().unwrap();
-            assert_eq!(after.created_at, expected_created_at);
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_touch_metadata_best_effort_preserves_corrupt_file() {
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_path = temp_dir.path().join("virtualenvs").join("broken");
-            let garbage = "{ this is not valid json";
-            seed_metadata_file(&env_path, garbage);
-
-            // Must not panic / propagate the error. Must not overwrite.
-            service.touch_metadata_at("broken", "2026-06-02T12:00:00Z".parse().unwrap());
-
-            let on_disk = fs::read_to_string(env_path.join(Metadata::FILE_NAME)).unwrap();
-            assert_eq!(
-                on_disk, garbage,
-                "corrupt metadata must be preserved verbatim — overwriting it \
-                 would destroy the user's only forensic trace of the corruption"
-            );
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_touch_metadata_best_effort_noop_on_missing_metadata() {
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_path = temp_dir.path().join("virtualenvs").join("nofile");
-            fs::create_dir_all(&env_path).unwrap();
-
-            // Legacy env with no metadata file. Must NOT synthesize one
-            // (that would lie about created_at) and must NOT error.
-            service.touch_metadata_at("nofile", "2026-06-02T12:00:00Z".parse().unwrap());
-            assert!(!env_path.join(Metadata::FILE_NAME).exists());
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_touch_metadata_best_effort_wall_clock_writes_recent() {
-        // Smoke test for the public no-arg entry point: it must actually
-        // call Utc::now() at write time, not silently no-op. We can't pin
-        // the exact value, so we assert it's bracketed by a before/after
-        // wall-clock window.
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_path = temp_dir.path().join("virtualenvs").join("wallclock");
-            let seed = r#"{
-                "name": "wallclock",
-                "python_version": "3.12",
-                "created_at": "2024-01-01T00:00:00Z",
-                "created_by": "scoop test",
-                "uv_version": null
-            }"#;
-            seed_metadata_file(&env_path, seed);
-
-            let before = Utc::now();
-            service.touch_metadata_best_effort("wallclock");
-            let after = Utc::now();
-
-            let m = service.read_metadata_result(&env_path).unwrap().unwrap();
-            let touched = m.last_used.expect("last_used set");
-            assert!(
-                touched >= before && touched <= after,
-                "{touched} not in [{before},{after}]"
-            );
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_read_metadata_legacy_api_swallows_corrupt() {
-        // The old `read_metadata` API returns `Option<Metadata>` and
-        // collapses both "missing" and "corrupt" into `None`. Pin that
-        // behavior explicitly so a mutant that returns `Some(default)`
-        // (won't compile — Metadata has no Default), or that flips the
-        // collapse, gets caught.
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_dir = temp_dir.path().join("virtualenvs").join("legacyapi");
-            fs::create_dir_all(&env_dir).unwrap();
-
-            // No file → None
-            assert!(service.read_metadata(&env_dir).is_none());
-
-            // Healthy file → Some
-            let seed = r#"{
-                "name": "legacyapi",
-                "python_version": "3.12",
-                "created_at": "2024-01-01T00:00:00Z",
-                "created_by": "scoop test",
-                "uv_version": null
-            }"#;
-            fs::write(env_dir.join(Metadata::FILE_NAME), seed).unwrap();
-            assert!(service.read_metadata(&env_dir).is_some());
-
-            // Corrupt file → None (collapses with missing, by design)
-            fs::write(env_dir.join(Metadata::FILE_NAME), "{ nope").unwrap();
-            assert!(service.read_metadata(&env_dir).is_none());
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_write_metadata_atomic_round_trips() {
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_path = temp_dir.path().join("virtualenvs").join("atomic");
-            fs::create_dir_all(&env_path).unwrap();
-
-            let mut meta = Metadata::new("atomic".to_string(), "3.12".to_string(), None);
-            let now = "2026-06-02T12:00:00Z".parse::<DateTime<Utc>>().unwrap();
-            meta.touch(now);
-
-            service.write_metadata_atomic(&env_path, &meta).unwrap();
-
-            let restored = service
-                .read_metadata_result(&env_path)
-                .unwrap()
-                .expect("file exists after atomic write");
-            assert_eq!(restored.last_used, Some(now));
-
-            // No tempfile left behind. tempfile::NamedTempFile::persist
-            // promises this but we assert it explicitly so any future
-            // change to the impl that breaks cleanup gets caught.
-            let leftover: Vec<_> = fs::read_dir(&env_path)
-                .unwrap()
-                .filter_map(|e| e.ok())
-                .filter(|e| {
-                    let n = e.file_name();
-                    let s = n.to_string_lossy();
-                    s.starts_with(".tmp") || s.contains("scoop-metadata.json.tmp")
-                })
-                .collect();
-            assert!(leftover.is_empty(), "tempfile residue: {leftover:?}");
-        });
-    }
-
-    #[test]
-    #[serial]
-    fn test_write_metadata_atomic_overwrites_existing() {
-        with_temp_scoop_home(|temp_dir| {
-            let service = require_uv!();
-            let env_path = temp_dir.path().join("virtualenvs").join("overwrite");
-            fs::create_dir_all(&env_path).unwrap();
-
-            // First write
-            let m1 = Metadata::new("overwrite".to_string(), "3.11".to_string(), None);
-            service.write_metadata_atomic(&env_path, &m1).unwrap();
-
-            // Second write replaces it atomically.
-            let mut m2 = Metadata::new("overwrite".to_string(), "3.12".to_string(), None);
-            m2.touch("2026-06-02T12:00:00Z".parse().unwrap());
-            service.write_metadata_atomic(&env_path, &m2).unwrap();
-
-            let on_disk = service.read_metadata_result(&env_path).unwrap().unwrap();
-            assert_eq!(on_disk.python_version, "3.12");
-            assert!(on_disk.last_used.is_some());
-        });
-    }
-
-    // C2 regression — symlinks under virtualenvs/ must NOT be enumerated.
-    // Otherwise downstream commands (gc, verify, ...) would treat the
-    // symlink target as a real env and end up scanning / exec'ing files
-    // outside the venvs dir.
-    #[cfg(unix)]
-    #[test]
-    #[serial]
-    fn test_list_skips_symlink_entries() {
-        with_temp_scoop_home(|temp_dir| {
-            let venvs_dir = temp_dir.path().join("virtualenvs");
-            fs::create_dir_all(&venvs_dir).unwrap();
-
-            // Real env so the list isn't empty (controls for "filter is
-            // entirely broken" vs "filter caught the symlink").
-            create_mock_venv(temp_dir, "real", None);
-
-            // Plant a symlink → some other (existing) directory. Without
-            // the symlink filter this would be enumerated as an env.
-            let other = tempfile::TempDir::new().unwrap();
-            std::os::unix::fs::symlink(other.path(), venvs_dir.join("symlinked")).unwrap();
-
-            let service = require_uv!();
-            let envs = service.list().unwrap();
-            assert_eq!(envs.len(), 1, "symlink entries must be skipped: {envs:?}");
-            assert_eq!(envs[0].name, "real");
-        });
-    }
-}
